+++ tests/web_tests/test_group.py
@@ -21,7 +21,10 @@ def test_group_with_class_func_negative(mouse_event_page):
     assert not mouse_click_card.drag_source_func().is_displayed()
 
 
-def test_group_object_in_all_elements(second_playground_page):
-    all_cards = second_playground_page.get_all_cards()
-    for element_object in all_cards:
-        assert 'WrappedCard' in str(element_object)
+def test_group_without_init(forms_page):
+    forms_page.validation_form.form_mixin.input.type_text('sample')
+    forms_page.validation_form.submit_form_button.click()
+    assert not forms_page.validation_form.invalid_city_error.is_displayed()
+    assert forms_page.validation_form.invalid_feedback_error.is_displayed()
+    assert forms_page.validation_form.invalid_zip_error.is_displayed()
+    assert forms_page.validation_form.invalid_terms_error.is_displayed()