+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 # Dyatel Changelog
 
 ---
+## v1.0.6
+*Release date: In development*
 
+---
 
 ## v1.0.5
-*Release date: In development*
+*Release date: 2022-10-07*
+
+### Added
+- `_first_element` property in `PlayElement`
+
+### Changed
+- `element` property replaced with `_first_element` for elements interactions
+- `parent` nesting of `Element` changed from one level to endless
+- `PlayElement` / `CoreElement` initialization
+
+### Fixed
+- `all_elements` execution time/nesting
 
 ---